move thumbSelector to theme
